git-svn-id: file:///semplestb/final_repo@960 02200ff9-b5b2-46f0-9171-221b09c08c7b

+++ SemplestServices/src/semplest/server/service/springjdbc/storedproc/AddAdvertisingEngineAccountSP.java
@@ -15,7 +15,7 @@ public class AddAdvertisingEngineAccountSP extends StoredProcedure
 	public AddAdvertisingEngineAccountSP()
 	{
 		super(BaseDB.jdbcTemplate.getDataSource(), SPROC_NAME);
-		declareParameter(new SqlParameter("AdvertisingEngineAccountID", Types.INTEGER));
+		declareParameter(new SqlParameter("AdvertisingEngineAccountID", Types.BIGINT));
 		declareParameter(new SqlParameter("AdvertisingEngine", Types.VARCHAR));
 		declareParameter(new SqlParameter("CustomerID", Types.INTEGER));
 		compile();
